Bug in dl test files. Introduced a bug where the test files were unzipped after DL breaking the default input file pattern.

+++ workflows/download_offline_files.py
@@ -73,8 +73,8 @@ def download_file(workflow, data, install_dir):
                             urllib.request.urlretrieve(url_string, install_dir+ "/"+ file_name, reporthook)
                         if (file_name.endswith('.tgz') or file_name.endswith('.tar')):
                             untar_command = "tar -zxvf " + install_dir+"/" + file_name + " -C " + install_dir + " && rm -f " + install_dir+"/" + file_name
-                            subprocess.run([untar_command], shell=True)
-                        elif (file_name.endswith('.gz')):
+                            subprocess.run([untar_command], shell=True) 
+                        elif (file_name.endswith('.gz') and not file_name.endswith('fq.gz')):
                             unzip_command = "gunzip -c " + install_dir+"/" + file_name + " > " + install_dir + "/" + os.path.splitext(file_name)[0] + " && rm -f " + install_dir+"/" + file_name
                             subprocess.run([unzip_command], shell=True)
                     except SocketError as e:
